requirements.txt
accelerate==0.29.3
fairscale==0.4.13
fire==0.6.0
grpcio
matplotlib==3.8.0
memory_profiler==0.61.0
networkx==2.6.3
numpy
pandas==1.5.3
Pillow==9.2.0
protobuf==3.20.3
pytorch_pretrained_bert==0.6.2
pyzmq==25.1.1
Requests==2.32.3
sentencepiece==0.1.99
setuptools==68.2.2
torch==2.3.0
torchvision
transformers==4.42.0
